fix: 修复 APK 启动失败 (LateInitializationError)
- Hive 初始化提到 runApp() 之前，避免并发初始化
- late final Box 改为可空类型，彻底避免重复初始化
- 加 _initializing 标志防止并发调用 init()
- initializationProvider 去掉 hive.init() 调用
- app_bootstrap.dart 只做数据完整性校验，不再调用 init()

+++ lib/domain/services/budget_service.dart
@@ -54,12 +54,12 @@ class BudgetService {
 
     final statuses = budgets.map((b) {
       final spent = spentByCat[b.categoryId] ?? 0;
-      final progress = b.limitAmount > 0 ? spent / b.limitAmount : 0;
+      final progress = b.limitAmount > 0 ? spent / b.limitAmount : 0.0;
       return BudgetStatus(
         budget: b,
         spent: spent,
         remaining: b.limitAmount - spent,
-        progress: progress,
+        progress: progress.toDouble(),
       );
     }).toList();
